Fix getting the new id for a post
Previously, in a list you could count the number of items from start
to end. This resulted in the new integer for a zero-based list.
E.g. items 0, 1, 2 meant length = 3. This is also the new id for the
blog post. However, this fails for a sorted set. So grab the last
entry and add one to the latest id.

For now, this will work. It is better although to create a redis
key which is incremented every time a new article is created.